+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 * The `Rails/HttpStatus` cop is unavailable if the `rack` gem cannot be loaded. ([@bquorning][])
+* Fix `Rails/HttpStatus` not working with custom HTTP status codes. ([@bquorning][])
 
 ## 1.23.0 (2018-02-23)